added more tests and moved them to a separate folder
relates to #68
