+++ react-ui/src/reducers/index.js
@@ -1,6 +1,5 @@
 import {combineReducers} from 'redux';
 
-import home from './home';
 import searchJobs from './search_jobs_reducer.js';
 import postJob from './post_job_reducer.js';
 import registerStudent from './register_student_reducer.js';
@@ -10,7 +9,6 @@ import applyJob from './apply_job.js';
 import fetchMyJobs from './my_jobs.js';
 
 export default combineReducers({
-    home,
     searchJobs,
     postJob,
     applyJob,
