+++ react-ui/src/reducers/search_jobs_reducer.js
@@ -23,7 +23,7 @@ export default (state = initialState, action) => {
     case SET_TERM:
         return {
             ...state,
-            term: action.response
+            searchTerm: action.response
         };
 
     case SET_SELECTED_JOB:
