+++ tests/reducers/search_students.test.js
@@ -1,5 +1,5 @@
-import reducer from '../search_students_reducer.js';
-import * as types from '../../constants/action_types.js';
+import reducer from '../../react-ui/src/reducers/search_students_reducer.js';
+import * as types from '../../react-ui/src/constants/action_types.js';
 
 describe('search students reducer', () => {
     it('should return the initial state', () => {
@@ -73,6 +73,20 @@ describe('search students reducer', () => {
                     }
                 ]
             }
+        });
+    });
+
+    it('should handle SET_TERM', () => {
+        expect(
+            reducer(
+                {},
+                {
+                    type: types.SET_TERM,
+                    response: 'coding'
+                }
+            )
+        ).toEqual({
+            searchTerm: 'coding'
         });
     });
 });
